feat: 리팩토링 및 새로운 컴포넌트 추가로 일정 관리 기능 개선
