+++ src/components/CalendarNavigation.tsx
@@ -1,0 +1,35 @@
+import { ChevronLeft, ChevronRight } from '@mui/icons-material';
+import { IconButton, MenuItem, Select, Stack } from '@mui/material';
+
+interface CalendarNavigationProps {
+  view: 'week' | 'month';
+  setView: (view: 'week' | 'month') => void;
+  navigate: (direction: 'prev' | 'next') => void;
+}
+
+export function CalendarNavigation({ view, setView, navigate }: CalendarNavigationProps) {
+  return (
+    <Stack direction="row" spacing={2} justifyContent="space-between" alignItems="center">
+      <IconButton aria-label="Previous" onClick={() => navigate('prev')}>
+        <ChevronLeft />
+      </IconButton>
+      <Select
+        size="small"
+        aria-label="뷰 타입 선택"
+        data-testid="view-selector"
+        value={view}
+        onChange={(e) => setView(e.target.value as 'week' | 'month')}
+      >
+        <MenuItem value="week" aria-label="week-option" data-testid="week-option">
+          Week
+        </MenuItem>
+        <MenuItem value="month" aria-label="month-option" data-testid="month-option">
+          Month
+        </MenuItem>
+      </Select>
+      <IconButton aria-label="Next" onClick={() => navigate('next')}>
+        <ChevronRight />
+      </IconButton>
+    </Stack>
+  );
+}
