+++ src/components/NotificationStack.tsx
@@ -1,0 +1,36 @@
+import { Close } from '@mui/icons-material';
+import { Alert, AlertTitle, IconButton, Stack } from '@mui/material';
+
+interface Notification {
+  message: string;
+}
+
+interface NotificationStackProps {
+  notifications: Notification[];
+  onClose: (index: number) => void;
+}
+
+export function NotificationStack({ notifications, onClose }: NotificationStackProps) {
+  if (notifications.length === 0) {
+    return null;
+  }
+
+  return (
+    <Stack position="fixed" top={16} right={16} spacing={2} alignItems="flex-end">
+      {notifications.map((notification, index) => (
+        <Alert
+          key={index}
+          severity="info"
+          sx={{ width: 'auto' }}
+          action={
+            <IconButton size="small" onClick={() => onClose(index)}>
+              <Close />
+            </IconButton>
+          }
+        >
+          <AlertTitle>{notification.message}</AlertTitle>
+        </Alert>
+      ))}
+    </Stack>
+  );
+}
